backend/requirements.txt
Flask==3.0.3
flask-cors==6.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pyserial==3.5
python-multipart==0.0.6
pydantic==2.5.0
asyncio-mqtt==0.16.1 
numpy
pandas 